remove cyclic dependencies
